+++ tests/step_test.py
@@ -7,7 +7,7 @@ from tango import StepGraph
 from tango.common import Params, Registrable
 from tango.common.from_params import FromParams
 from tango.common.testing import TangoTestCase
-from tango.step import Step
+from tango.step import FunctionalStep, Step, step
 
 
 class TestStep(TangoTestCase):
@@ -44,6 +44,23 @@ class TestStep(TangoTestCase):
         step1 = SkipArgStep(arg="foo")
         step2 = SkipArgStep(arg="bar")
         assert step1.unique_id == step2.unique_id
+
+    def test_skip_default_arguments(self):
+        class SkipArgStep(Step):
+            def run(self) -> int:  # type: ignore
+                return 5
+
+        old_hash = SkipArgStep().unique_id
+
+        class SkipArgStep(Step):
+            SKIP_DEFAULT_ARGUMENTS = {"arg": 5}
+
+            def run(self, arg: int = 5) -> int:  # type: ignore
+                return arg
+
+        assert SkipArgStep().unique_id == old_hash
+        assert SkipArgStep(arg=5).unique_id == old_hash
+        assert SkipArgStep(arg=6).unique_id != old_hash
 
     def test_massage_kwargs(self):
         class CountLettersStep(Step):
@@ -134,3 +151,19 @@ class TestStep(TangoTestCase):
         }
         sg = StepGraph.from_params(config)
         assert len(sg["holder_consumer"].dependencies) > 0
+
+    def test_functional_step(self):
+        class Bar(FromParams):
+            def __init__(self, x: int):
+                self.x = x
+
+        @step(exist_ok=True)
+        def foo(bar: Bar) -> int:
+            return bar.x
+
+        assert issubclass(foo, FunctionalStep)
+        assert foo().run(Bar(x=1)) == 1
+
+        foo_step = Step.from_params({"type": "foo", "bar": {"x": 1}})
+        assert isinstance(foo_step, FunctionalStep)
+        assert isinstance(foo_step.kwargs["bar"], Bar)